refactor: move cli export helpers
