--- .github/max-lines-ignore
@@ -1,1 +1,0 @@
-docling/cli/main.py
